qor_core::component
Trait Component

pub trait Component {
    // Required methods
    fn init(&mut self) -> CoreResult<State>;
    fn setup(&mut self, config: &Config) -> CoreResult<State>;
    fn start_up(&mut self) -> CoreResult<State>;
    fn run(&mut self) -> CoreResult<State>;
    fn shut_down(&mut self) -> CoreResult<State>;
    fn class_name() -> &'static str;
    fn instance_name(&self) -> &str;
    fn state(&self) -> State;
}

A component is an element with a simple life cycle

Required Methods

fn init(&mut self) -> CoreResult<State>

Initialize the component After this the component is ready for setup

fn setup(&mut self, config: &Config) -> CoreResult<State>

Setup the component The component will continue initialization

fn start_up(&mut self) -> CoreResult<State>

fn run(&mut self) -> CoreResult<State>

fn shut_down(&mut self) -> CoreResult<State>

fn class_name() -> &'static str

fn instance_name(&self) -> &str

fn state(&self) -> State

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.