qor_core
Module types

Structs

StrConst

Constant string type

TypeId

Compiler-independent type IDs for built-in types

bf16

The brain float 16 type

f16

The half-precision IEEE754 float 16 type

Traits

Coherent

The coherent trait marks a type as coherent in memory.

Floating

The Coherent trait declares a type as float

Inf

Trait that returns the Inf value of a type Inf is the infinity element of that type

Integral

The Integral trait declares a type as integral

Lockfree

The Lockfree marker trait defines a type as atomically lockfree accessible.

One

Trait that returns the One value of a type One is defined as the neutral element of multiplication

Reloc

The Reloc marker trait defines a type as relocatable in memory.

Scalar

The Scalar trait declares a type as scalar

TypeTag

Indicator trait that a type supports compiler-independent type tags.

Zero

Trait that returns the ZERO value of a type Zero is defined as the neutral element of addition