qor_core::types
Trait Coherent

pub trait Coherent { }

The coherent trait marks a type as coherent in memory.

Coherent means that all data of the type is stored in adjacent memory locations (a range) and no data of the type is stored outside this range.

All Scalar types are Coherent: bool, u8, i8, u16, i16, .. , u128, i128, f32, f64, char Most Rust types like std::String and &str are not Coherent.

Container types of the standard library are also not coherent. The qor_data crate provides coherent implementations of collections.