qor_core::types
Trait Lockfree

pub unsafe trait Lockfree: Sync { }

The Lockfree marker trait defines a type as atomically lockfree accessible.

Lockfree is an extension of Sync and a stronger statement, because a lockfree type is not only free from inner mutability, it also guarantees to not lock the accessing thread during access. Lockfree does not imply Nonblocking. Lockfree types may use short spinlock to prevent data races.

Types marked as lockfree can be accessed without acquiring a lock.

Scalar types up to 64bits can safely be assumed Lockfree on 64bit architectures.

Safety

This trait has no safety implications and is only a marker.