qor_core::types
Trait Reloc

pub unsafe trait Reloc { }

The Reloc marker trait defines a type as relocatable in memory.

Relocatable types do not contain pointers or references. They are well defined independent of the memory space they are located in. The Relocatable trait is essential for data stored in shared memory locations.

All Scalar types are Reloc: bool, u8, i8, u16, i16, .. , u128, i128, f32, f64, char Most other Rust types, especially those that allow dynamic length data or are generics are not Reloc: &str, String, Vec, HashMap, Box, Rc, Arc, Mutex, …

Safety

This trait has no safety implications and is only a marker.